+++ tests/unit/test_cloudwanderer_resource.py
@@ -68,6 +68,7 @@ class TestCloudWandererResource(unittest.TestCase):
             "CloudWandererResource("
             "urn=URN(account_id='111111111111', region='eu-west-2', service='ec2', "
             "resource_type='vpc', resource_id='vpc-11111111'), "
+            "parent_urn=None, "
             "resource_data={'CidrBlock': '10.0.0.0/0'}, secondary_attributes=[{'EnableDnsSupport': {'Value': True}}]"
             ")"
         )
@@ -83,6 +84,7 @@ class TestCloudWandererResource(unittest.TestCase):
             "CloudWandererResource("
             "urn=URN(account_id='111111111111', region='eu-west-2', service='ec2', "
             "resource_type='vpc', resource_id='vpc-11111111'), "
+            "parent_urn=None, "
             "resource_data={'CidrBlock': '10.0.0.0/0'}, secondary_attributes=[{'EnableDnsSupport': {'Value': True}}]"
             ")"
         )
@@ -98,3 +100,10 @@ class TestCloudWandererResource(unittest.TestCase):
             resource_data=None,
             secondary_attributes=[],
         )
+
+    def test_subresource(self):
+        parent_urn = URN.from_string("urn:aws:111111111111:us-east-1:iam:role:test-role")
+        urn = URN.from_string("urn:aws:111111111111:us-east-1:iam:role_policy:test-role/test-policy")
+        cwr = CloudWandererResource(urn=urn, parent_urn=parent_urn, resource_data={}, secondary_attributes=[])
+
+        assert cwr.parent_urn == parent_urn